Refactoring the folders
