+++ api/server.py
@@ -1,7 +1,7 @@
-from api import app
+from app import app
 
-from api.routes.ingredient import ingredient_index, ingredient_category, ingredient_all, ingredient_create
-from api.routes.recipe import recipe_index, recipe_all, recipe_create
+from app.routes.ingredient import ingredient_index, ingredient_category, ingredient_all, ingredient_create
+from app.routes.recipe import recipe_index, recipe_all, recipe_create
 
 if __name__ == '__main__':
     app.run(debug=True, use_reloader=False)
